[hotfix] move the model mapper for Gaussian Mixture model to common
