+++ flink-ml-parent/flink-ml-lib/src/main/java/org/apache/flink/ml/common/ml/clustering/GmmModelMapper.java
@@ -25,10 +25,10 @@ import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.ml.api.misc.param.Params;
 import org.apache.flink.ml.common.linalg.Vector;
 import org.apache.flink.ml.common.mapper.RichModelMapper;
+import org.apache.flink.ml.common.statistics.basicstatistic.MultivariateGaussian;
 import org.apache.flink.ml.common.utils.JsonConverter;
 import org.apache.flink.ml.common.utils.TableUtil;
 import org.apache.flink.ml.params.ml.clustering.GmmPredictParams;
-import org.apache.flink.ml.statistics.basicstatistic.MultivariateGaussian;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.types.Row;
